2_vazifa_ML
Savol:

Sklearn.datasets:
                   diabetes = datasets.load_diabetes()

                   X= diabetes.data[:, np.newaxis, 2]       # (442, 1)  행렬   <-> (442,)
                   X2 = diabetes.data: [:,0:3]                     # (442, 3) 행렬
                   y = diabetes.target                                                                                                                    b. Solve the similar problem:
-
-
-                                                                                                                                                                                                             -

-  x' = np.array(diabetes.data[23, 2])   -> y' = ?
-  x''= diabetes.data[23, 0:3]                -> y" = ?

Javob:

The load_diabetes function from the sklearn.datasets module loads the diabetes dataset, which is a sample dataset for regression analysis.

The dataset contains 10 features for 442 patients. X is a matrix with shape (442, 1) that contains only the third feature of the original dataset, and X2 is a matrix with shape (442, 3) that contains the first three features of the original dataset. y is a vector with shape (442,) that contains the target

Given thex' = np.array(diabetes.data[23, 2]),x' as a 2D arraypredict method

Given thex'' = diabetes.data[23, 0:3],x'' as a 2D arraypredict method